Create enriched JSON-LD for elasticsearch using N-Triple dumps only
This replaces the experimental approach using our triple store
(which was slow in the first place and wouldn't scale w/ the cluster).

The basic idea is to use two Hadoop jobs: the first job creates the
resolved/enriched triples, while the second job converts the triples
to JSON-LD (as it did originally). The first job uses jfresnel and
a Fresnel Selector Language (FSL) expression to resolve creators.

For some reason, this new approach is actually a little faster than
the original JSON-LD conversion (which did not resolve anything).

The included MRUnit tests need multiple test inputs, which is only
supported in MRUnit 1.0.0, so this also adds a dependency to MRUnit
SNAPSHOT releases (which have to be installed into the local repo).