Configure all-in-one Jar dependency for Metatext Eclipse bundle
With `mvn clean assembly:assembly` a Jar containing all lodmill-rd
dependencies is created, which is used by the Metatext Eclipse
bundle. The properties files currently don't end up in the
resulting Jar (they are overwritten by others in the dependencies).
They need to be copied manually into the Jar to make the Metaflow
workflows run when executed with Metatext in Eclipse.