Resolve references to the current directory ('.') in *.flow files
The default Metaflow support for portable *.flow files does not
work when running in an Eclipse plugin on a workspace file
('current' directory when running is not where the input file is).

This commit adds support for portable *.flow files with direct
execution in Eclipse. When running a *.flow file in the workspace,
read its content and replace '.' references with the absolute path
to the workspace location containing the *.flow file to run. Write
the new content to a temp file and pass that temp file to Metaflow.