Initial API for searching lobid-resources via ID
This commit adds API to search for lobid-resources based on IDs
currently present in our data: lobid-HT numbers, ISBN10, ISBN13.

To avoid conflicts with the existing author search (which does
some fancy stuff to allow search queries with life dates), this
also introduces a new parameter ('category') to specify where to
search ('author' vs. 'id').

The UI providing documentation and sample usage of the API is
updated and allows for the selection of indexes and their
corresponding categories.