WIP: basic approach for collecting triples with distributed cache
This breaks some unit tests which are commented out and it doesn't
yield the desired JSON-LD yet (it seems like blank nodes are not
handled correctly in Jena or jsonld-jena).

Also, this would not scale as implemented here - we would at least
need to use map files for the cached subjects to avoid keeping them
in RAM on every node.