Replace new `field` param with `format=short` modifier (see #230)
With full parameters, the new syntax was counter-intuitive, e.g.
`GET /resource?id=XY&format=full&field=publisher` does not return
the full resource, but only the values of the `publisher` field.

Since the new result format is syntactically identical to the
existing `short` format, it seems to make sense to use a modifier
on that instead, e.g. `GET /resource?id=XY&format=short.publisher`.

It also makes more sense semantically, since the field is not a
property of the query, but of the result format. Finally, it also
points users to the option of using this for custom auto-suggest
values (which is what the default `short` format is intended for).