Don't identify primary topic in external JSON-LD, only internally
Identifying the primary topic was originally both for the request
in 230#issuecomment-32290484 and for #257. It did break Edoweb
however, probably due to an issue with parsing named JSON-LD graphs
in rdflib. This removes the primary topic from the served JSON-LD,
but retains it in the internal format, used by the UI for #257.