Outcomment provenance and about URIs
See #541.

The about URIs are already created in the API with "primaryTopicOf"
added to them.
The triples of the about URIs generated in transformation
are discarded in API anyway, so the <void:inDataset> triple was never shown.
As to achieve daily updates we discard about URIs and provenance.
May be added later dynamically in the API. That's easy, because
the information to be added is of static nature.

* update tests