Add RdfModel2ElasticsearchJsonLd
See hbz/lobid#139.

* add test and test data

This metafacture module generates json-ld from a jena rdf model.
The generated documents are ready to be elasticsearch bulk indexed.

There are highly specific requirements for generating documents out of the
hbz01 catalog graph build with morph. One hbz01 catalog entry may result in
dozens of documents, namely the

- 'main' doc (data about the main resource)
- 'items' of a doc
- 'super' docs ("hasPart" nodes)
- 'sameAs' docs

But not all nodes should be nodes on their own: the gnd nodes must stay
sub nodes of the main node.

So this module is not generic and may be made generic only to a certain degree.