Add runner; add script helping check test differences; fix bulk indexing
As we don't want to make use of metafacture flow anymore a "run" package
is added for starting processes. The "flow" class starts the
transformation and json conversion and indexing into elasticsearch.

Fixed: The bulk indexer was not reset so that update requests were ever more added
and indexed all over again which which results in low performance of course.

* update tests

See hbz/lobid#139.