Fix buggy index-config.json
We had some redundancy observed in the index mappings of elasticsearch.
Most things run smoothly enough and we thought that that was ok. But it is not.

This commit adjusts the config mappings so that a lookup of the mappings of the
index dont't hold any "redundancy" (more accurate: "not used definitions").
It should fix the type-query mentioned in hbz/lobid#139 and also hbz/lobid#141
(see the commit a6219bbb3bf5596b3a030da1e489fc1ba852d60a "... the @graph.@id of
the items wasn't used yet" ).