Remove building lodmill-ld
As of hbz/lobid#139 we (mostly) don't use lodmill-ld anymore.
It's jsut needed by the old lobid-organisations , which will be
exchanged with a new way to make the data. Furthermore, the old
lobid-organisations will not be enhanced anymore. Thus, it is expected
to not alter lodmill-ld anymore. Thus, it is safe to remove lodmill-ld
form the build processes, especially for travis since such a build with
all the tests and mockups take around 7 minutes or so and even sometimes
fail because travis has problems with it (memory, especially).