Add checkout and building for every transformation in the loop
It _could_ be that, while a transformation is triggered in the loop, another
build process of another branch changed the jar. So, if the next transformation
in the loop is triggered there would be a false jar. So it's safer to rebuild
the jar every time a transformation is started.
Note:
It may be that two builds are started in parallel. This would break the building.
But this breaking would be more overt (and the chances that they occur are smaller)
than to use a build from a totally different branch.